# src/backend/api_spec.yaml
openapi: 3.0.1
info:
  title: Sayyara API
  description: API Explorer for Sayyara
  termsOfService: https://localhost/
  contact:
    email: example@example.com
  license:
    name: Closed License
  version: v1
servers:
  - url: http://localhost/api
security:
  - Bearer: []
paths:
  /account/{id}:
    get:
      tags:
        - Profile
      description: Returns the profile details of an account
      operationId: account_read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content: {}
    put:
      tags:
        - Profile
      description: Update a profile
      operationId: account_update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileUpdate'
      x-codegen-request-body-name: data
...
components:
  schemas:
    ProfileUpdate:
      required:
        - email
        - first_name
        - last_name
      type: object
      properties:
        first_name:
          title: First name
          minLength: 1
          type: string
        last_name:
          title: Last name
          minLength: 1
          type: string
        email:
          title: Email
          minLength: 1
          type: string
        phone:
          title: Phone
          maxLength: 15
          minLength: 1
          pattern: ^\d{3}-\d{3}-\d{4}$
          type: string
          nullable: true
        address:
          title: Address
          maxLength: 200
          minLength: 1
          type: string
          nullable: true
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header